Passage des sondes en long long int pour éviter les dépassements

+++ src/sorts/probe.h
@@ -4,4 +4,4 @@
 
 #pragma once
 
-int __COMPARISION_COUNT__;
+long long int __COMPARISION_COUNT__;
